requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
requests
pydantic
